+++ stochss_compute/server/is_cached.py
@@ -20,7 +20,7 @@ stochss_compute.server.is_cached
 from datetime import datetime
 from tornado.web import RequestHandler
 from stochss_compute.core.errors import RemoteSimulationError
-from stochss_compute.core.messages import SimStatus, StatusResponse
+from stochss_compute.core.messages.status import SimStatus, StatusResponse
 from stochss_compute.server.cache import Cache
 
 class IsCachedHandler(RequestHandler):
