+++ stochss_compute/server/run.py
@@ -25,7 +25,8 @@ from tornado.web import RequestHandler
 from tornado.ioloop import IOLoop
 from distributed import Client, Future
 from gillespy2.core import Results
-from stochss_compute.core.messages import SimStatus, SimulationRunRequest, SimulationRunResponse
+from stochss_compute.core.messages.status import SimStatus
+from stochss_compute.core.messages.simulation_run import SimulationRunRequest, SimulationRunResponse
 from stochss_compute.server.cache import Cache
