+++ test/unit_tests/test_status_handler.py
@@ -5,7 +5,9 @@ import os
 import subprocess
 from test.unit_tests.gillespy2_models import create_michaelis_menten
 from tornado.testing import AsyncHTTPTestCase
-from stochss_compute.core.messages import SimStatus, SimulationRunRequest, StatusResponse
+from stochss_compute.core.messages.base import SimStatus
+from stochss_compute.core.messages.simulation_run import SimulationRunRequest
+from stochss_compute.core.messages.status import StatusResponse
 from stochss_compute.server.api import _make_app
 from stochss_compute.server.cache import Cache
